+++ ContentPatcher/release-notes.md
@@ -1,7 +1,8 @@
 [← back to readme](README.md)
 
 # Release notes
-## 1.3.1
+## 1.4
+* Added more info to `patch summary` command.
 * Improved error-handling when applying image edits.
 * Fixed unnecessary warnings when a patch has `Enabled: false`.
 * Fixed error when a content pack's `config.json` has invalid keys.